отрефакторены main.js и films-list-presentor.js под отдельный презентер для карточки фильма
